Added first method to BookmarkController and solved Json Recursion Mapping Exception

+++ bookmark-rest/src/test/java/com/myprojects/bookmarkrest/dao/BookmarkDaoTest.java
@@ -13,6 +13,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 
+import java.util.List;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+
 @RunWith(SpringRunner.class)
 @SpringBootTest(classes = BookmarkRestApplication.class)
 public class BookmarkDaoTest
@@ -32,14 +37,14 @@ public class BookmarkDaoTest
         testFolder = new Folder();
         testFolder.setName("Test");
 
-        folderDao.create(testFolder);
+        folderDao.save(testFolder);
     }
 
     @After
     public void tearDown() throws Exception
     {
-        bookmarkDao.delete(testBookmark);
-        folderDao.delete(testFolder);
+        if (testBookmark != null) bookmarkDao.delete(testBookmark);
+        if (testFolder != null) folderDao.delete(testFolder);
     }
 
     @Test
@@ -50,6 +55,32 @@ public class BookmarkDaoTest
         testBookmark.setUrl("http://www.google.com");
         testBookmark.setFolder(testFolder);
 
-        bookmarkDao.create(testBookmark);
+        bookmarkDao.save(testBookmark);
+    }
+
+    @Test
+    public void testFindByFolderIdPositive() {
+        String description = "Test description";
+        String url = "http://www.google.com";
+
+        testBookmark = new Bookmark();
+        testBookmark.setDesc(description);
+        testBookmark.setUrl(url);
+        testBookmark.setFolder(testFolder);
+
+        bookmarkDao.save(testBookmark);
+
+        List<Bookmark> bookmarks = bookmarkDao.findByFolderFolderId(testFolder.getFolderId());
+        assertNotNull(bookmarks);
+        assertEquals(1, bookmarks.size());
+        assertEquals(description, bookmarks.get(0).getDesc());
+        assertEquals(url, bookmarks.get(0).getUrl());
+    }
+
+    @Test
+    public void testFindByFolderIdNegative() {
+        List<Bookmark> bookmarks = bookmarkDao.findByFolderFolderId(0);
+        assertNotNull(bookmarks);
+        assertEquals(0, bookmarks.size());
     }
 }
